Convert InstanceType to a CSV
This is super ugly internally, but it presents a possibly OK interface
to users of the stack.

Users can specify 1-4 instance types separated by commas. When
launching OnDemand instances, the types will be attempted in priority
order. When launching spot instances (if they're enabled) all types are
considered.

Internally, this is implemented by appending 3 commas to the
InstanceType and re-splitting it (to ensure there's ALWAYS at least 4
list items passed to !Select.